online_boutique/pyproject.toml
[tool.poetry]
name = "online-boutique"
version = "0.1"
description = "AI-driven multi-agent e-commerce system designed to provide comprehensive online shopping experience"
authors = ["Online Boutique Team"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "online_boutique_manager" }]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-aiplatform = { version = "^1.93.0", extras = [
    "adk",
    "agent-engines",
] }
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-adk = "^1.0.0"
flask = "^3.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
google-adk = { version = "^1.0.0", extras = ["eval"] }
pytest-asyncio = "^0.26.0"

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
